Bug fixes
Settings would:
a) Set the user to not attending if it was not roll call time.
b) After the page reloads from a settings update, the new settings were
not displayed...and settings.php would error with a "User settings not
found". I think the session may not be started in this case.