Fixes #52
This was actually cause by an unnoticed php error. $_POST['rollCall'] is
not necessarily set. I tried to fix this earlier by having it in the
form either way, but disabled, but this way is better. Also, should have
been "yes" not "Yes", which is probably why the previous fix didn't seem
to help very much.